Move most of the testing logic to Spring Boot 2.0.x app
