+++ dev/com.ibm.ws.springboot.support.version15.test.app/build.gradle
@@ -11,8 +11,6 @@ sourceCompatibility = 1.8
 
 dependencies {
 	compile('org.springframework.boot:spring-boot-starter-web')
-	compile('org.springframework.boot:spring-boot-starter-thymeleaf')
-	compile('org.springframework.boot:spring-boot-starter-security')
 	compile 'org.webjars:bootstrap:3.3.1'
 	compile 'org.webjars:webjars-locator:0.33'
 }
