+++ dev/com.ibm.ws.springboot.support.version20.test.app/src/main/resources/templates/hello.html
